Allow `Alt` as global keybinding
This change adds the option of (and defaults to) using `Alt_L` as the
global keybinding for activating mate-hud.

For this to work, there is a fix for Marco that allows it to process
synthetic events forwarded from other clients. Otherwise Marco swallows
any `<Alt>*` bindings (as well as `<Alt>MouseButtonPress` actions):
https://github.com/mate-desktop/marco/pull/342