Send current time when ungrabbing (#7)
Seems like a different process is creating an XEvent object in the queue
without the `time` attribue. python-xlib raises an exception when an
XEvent attribute is missing, which causes mate-hud to crash.

This fix sends `X.CurrentTime` instead to make sure that the X server
uses the latest place in the queue, this in turn updates the last event
to use `X.CurrentTime` which makes python-xlib happy.