Recover from unintended Alt grabs meant for a window
Determine which keycodes are supported as window keybindings by the
active display, and Replay the events when recovering from an unintended
keyboard grab. This lets the focused window grab the corresponding event
to operate on itself. Otherwise, send to asynchronously to the top-level
window, which allows the settings-daemon or window manager to deal with
it appropriately.

More concretely, it fixes the issue in
https://ubuntu-mate.community/t/how-to-fix-broken-alt-key/14602/51
so that the user does not need to press a modified key twice for the
window to capture the event.